requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.11.0
bce-python-sdk==0.9.46
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
colorlog==6.10.1
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.19.1
fsspec==2025.10.0
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
idna==3.11
imagesize==1.4.1
imutils==0.5.4
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.78
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
lxml==6.0.2
modelscope==1.31.0
networkx==3.2.1
numpy==2.0.2
openai==2.7.0
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
prettytable==3.16.0
protobuf==6.33.0
psutil==7.1.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pypdfium2==5.0.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
receipt-ocr==0.3.1
requests==2.32.5
requests-toolbelt==1.0.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scipy==1.13.1
shapely==1.8.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.14
xlsxwriter==3.2.9
xxhash==3.6.0
zstandard==0.25.0